[analysis_server] Ensure errors are cleared from NotificationManager cache when flushed on client upon file remove
When a server plugins generates diagnostics, they're merged with the servers diagnostics (which are kept cached in `NotificationManager.errors`) and sent to the client.

When a file is removed, we tell the client to flush all results but we did not previously clear this cache. That meant it a plugin subsequently generates diagnostics (even if empty), we would merge with them with the last state of the server errors and send them to the client.

This clears errors from the servers cache at the same time as flushing them on the client.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5113

Change-Id: I047c4531a38edaaad1362f9c7e7714a7e7ab705f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/367321
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>