[vm] Avoid truncated initial definitions in FlowGraphPrinter.
FlowGraphPrinter::PrintOneInstruction uses a stack-allocated
statically-sized buffer to print the instruction. If the instruction is
a block entry with initial definitions, then that buffer would need to
be large enough to hold the printed information for the block entry and
all of its initial definitions. However, in some pathological cases,
like the Dart fuzzer, graphs are created where the initial definitions
can end up truncated due to the use of a large number of constants.

Instead, add the following protected methods to
BlockEntryWithInitialDefs:

* PrintBlockHeaderTo, a pure virtual method that specifies how to print
  the block header information, overridden in each subclass.

* PrintInitialDefinitionsTo, a non-virtual method that prints the
  initial definitions. It takes an additional argument, a callback which
  takes the buffer being used for printing, and invokes the callback
  after printing each initial definition.

Also adds BlockEntryWithInitialDefs::PrintTo, which calls
PrintBlockHeaderTo and PrintInitialDefinitionsTo in sequence, providing
a no-op callback to the latter, and removes the PrintTo overrides in
the subclasses in favor of this base implementation.

Adjusts FlowGraphPrinter::PrintOneInstruction to call PrintBlockHeaderTo
and PrintInitialDefinitionsTo for BlockEntryWithInitialDefs
instructions, providing a callback that prints the current contents of
the buffer and then clears it. With this, it is unlikely using a
statically-sized stack-allocated buffer as an intermediate will cause
truncation, as each initial definition usually fits on a single line.

TEST=ci (internal debugging-only output, tested manually)

Issue: https://github.com/dart-lang/sdk/issues/55864
Change-Id: I7a7919cd6d76a22e018603240fe0ba6a09c2e6d3
Cq-Include-Trybots: luci.dart.try:vm-aot-mac-release-arm64-try,vm-aot-linux-debug-simarm_x64-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/368580
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Tess Strickland <sstrickl@google.com>