Store the static type on TypeState
This allows users to access the static type of a DartObject through `DartObject#toStaticTypeValue`.

I'm not sure whether this is the best approach or whether it'll solve all future use-cases. I suspect it might be nicer to instead have the TypeState maintain the static type as the default "type value" and have the analyzer erase the types for comparison when necessary. At least as a maintainer of Angular, I feel like I only really care about the static type while the runtime type isn't particularly useful to me.

Please let me know if you think there's a better way to address https://github.com/dart-lang/sdk/issues/55693 or if you think this won't cover it.

For an example in use, see cl/639257690.