analyzer: Move DartLinter to linter package's 'test_utilities'
And rename it to TestLinter.

We also get to remove `LintRule._locationInfo`,which was _only_ used
for testing pubspec-oriented rules.

This DartLinter class is only used in tests and tools of the linter
package, so it's good to move it out of analyzer lib/.

In order to remove `LintRule._locationInfo`, we move the singular
package_names test case to be a reflective test.

Cq-Include-Trybots: luci.dart.try:flutter-analyze-try,analyzer-win-release-try,pkg-win-release-try
Change-Id: I97b5e79daa07eb3942f5444502473a4ec1e9daf0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/369562
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>