Add repro instructions to coverage test output.
Also update the coverage test so that it prints detailed results to
standard output, so that the repro instructions will produce a useful
result.

In my experiments, I've found that the number supplied for the
`--tasks` argument can have a small effect on coverage numbers (though
I'm not sure why), so I've included it in the repro instructions.

Change-Id: Ic347a9e5e3c2da0f73ceefbaa9322a60295735d6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/368920
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>