[stable] [dart2wasm] Make `dart compile wasm` compiled apps disable `dart.library.ffi`
This is a follow-up to [0]. That CL changed dart2wasm's modular
transformer to issue an error if `dart:ffi` is imported.

Users of packages that have specialized code for the VM (which supports
FFI) use conditional imports based on `dart.library.ffi`. We don't want
the VM-specific code to be used for web in dart2wasm (as dart2wasm
doesn't support the entirety of `dart:ffi`).

As a result we're going to make `dart.library.ffi` be false in
coditional imports (as well as in
`const bool.fromEnvironment('dart.library.ffi')`).

[0] https://dart-review.googlesource.com/c/sdk/+/368568

Bug: https://github.com/dart-lang/sdk/issues/55948
Bug: https://github.com/flutter/flutter/issues/149984

Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/370580
Cherry-pick-request: https://github.com/dart-lang/sdk/issues/55979
Change-Id: Ia968bacf92566d421606fc026c7f016fe5abff01
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/370880
Reviewed-by: Slava Egorov <vegorov@google.com>